virtualQC fails after SO update

+++ bin/review_vqcfs.pl
@@ -52,9 +52,12 @@ while (@ARGV and $ARGV[0] =~ /^-/) {
 }
 
 # Mira, hay que meter el proyecto de XNAT con alguno de los dos switch
-$xprj = $prj unless $xprj;
+if ($prj and not $xprj) {
+	%pdata = load_project($prj);
+	$xprj = %pdata{'XNAME'};
+}
 # O te vas a tomar por culo
-die unless $xprj;
+die "Should supply XNAT project name or define it at local project config!\n" unless $xprj;
 # You must be under X11 to run this shit
 die "Should run this under X!\n" unless $ENV{'DISPLAY'};
 
@@ -67,6 +70,9 @@ my %xconf = xget_session();
 my @pollos;
 my %epollos;
 $efile=$wdir.'/'.$xprj.'_experiment.list';
+
+#Añadir logging!!!!!
+
 if ($ifile and -f $ifile){
 	$efile=$wdir.'/'.$xprj.'_experiment_custom.list';
 	open IDF, "<$ifile";
@@ -114,11 +120,15 @@ unless ($odir and -d $odir) {
 			$epollos{$pollo} = xget_mri($xconf{'HOST'}, $xconf{'JSESSION'}, $xprj, $pollo);
 			my $otfile = $tsdir.'/'.$pollo.'.tar.gz';
 			xget_fs_data($xconf{'HOST'}, $xconf{'JSESSION'}, $xprj, $epollos{$pollo}, $otfile);
-			my $otdir = $odir.'/'.$epollos{$pollo};
-			mkdir $otdir;
-			my $pre = 'tar xzf '.$otfile.' --strip-components=1 -C '.$otdir.' */mri/{orig,aparc+aseg}.mgz */label/*.aparc.annot */surf/*.pial* */stats/*.aparc.stats';
-			system($pre);
-			unlink $otfile;
+			if (-e $otfile) { 
+				my $otdir = $odir.'/'.$epollos{$pollo};
+				mkdir $otdir;
+				my $pre = 'tar xzf '.$otfile.' --strip-components=1 -C '.$otdir.' */mri/{orig,aparc+aseg}.mgz */label/*.aparc.annot */surf/*.pial* */stats/*.aparc.stats';
+				system($pre);
+				unlink $otfile;
+			} else {
+				print "No FS data for $pollo!\n";
+			}
 		}
 		unlink $tsdir;
 	}
